+++ include/Stream.h
@@ -32,15 +32,15 @@ public:
 	virtual ~Stream();
 
 public:
-	void Bind(unsigned char *pBuffer, unsigned uSize);//绑定一个缓冲区，在此之上做转换
+	void Bind(unsigned char *pBuffer, unsigned int uSize);//绑定一个缓冲区，在此之上做转换
 	//取得缓冲容量(对应创建数据流)/消息长度(对应解析数据流)
-	unsigned short GetSize();
+	unsigned int GetSize();
 //////////////////////////////////////////////////////////////////////////
 //组装消息流，用于输出
 	//取得组装好的数据流,并在头部设置数据长度
 	unsigned char* GetStream();
 	//取得当前游标位置（对于组装消息即数据长度，对于解析消息即下一个数据的开始位置）
-	unsigned short Pos();
+	unsigned int Pos();
 	//组装方法，填充参数到流
 	/**
 	 * 增加1个参数到消息中,char、unsigned char、int等各种类型的重载
@@ -150,9 +150,9 @@ protected:
 	//绑定的缓冲区
 	unsigned char* m_pMsgBuffer;
 	//缓冲大小/被解析消息长度
-	unsigned short m_uSize;
+	unsigned int m_uSize;
 	//消息填充/解析到的位置
-	unsigned short m_uPos;
+	unsigned int m_uPos;
 	
 };
